lambda_python/requirements.txt
boto3>=1.34.0
Pillow>=10.0.0
pydantic>=2.5.0
hypothesis>=6.92.0
pytest>=7.4.0
